requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu128

torch==2.10.0
torchvision==0.25.0
torchaudio==2.10.0
transformers==5.3.0
accelerate==1.13.0
datasets>=3.0
peft>=0.13
trl>=0.12
diffusers
einops
safetensors
evaluate
wandb
hydra-core
numpy
pandas
scipy
scikit-learn
huggingface_hub
pyyaml
